Refactored the code by adding new widget
