+++ collect_app/src/main/java/org/odk/collect/android/widgets/SelectOneWidget.java
@@ -42,7 +42,7 @@ import java.util.List;
  */
 @SuppressLint("ViewConstructor")
 public class SelectOneWidget
-        extends SelectWidget
+        extends TextSelectWidget
         implements OnCheckedChangeListener, AudioPlayListener, MultiChoiceWidget {
 
     protected List<RadioButton> buttons;
