+++ collect_app/src/main/java/org/odk/collect/android/widgets/SelectWidget.java
@@ -18,17 +18,9 @@ package org.odk.collect.android.widgets;
 
 import android.content.Context;
 import android.media.MediaPlayer;
-import android.text.Editable;
-import android.text.Selection;
-import android.text.TextWatcher;
-import android.util.TypedValue;
-import android.view.ViewGroup;
 import android.view.inputmethod.InputMethodManager;
-import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.RelativeLayout;
-import android.widget.TableLayout;
 import android.widget.TextView;
 
 import org.javarosa.core.model.SelectChoice;
@@ -38,20 +30,15 @@ import org.javarosa.form.api.FormEntryPrompt;
 import org.javarosa.xpath.expr.XPathFuncExpr;
 import org.odk.collect.android.external.ExternalDataUtil;
 import org.odk.collect.android.external.ExternalSelectChoice;
-import org.odk.collect.android.utilities.ViewIds;
 import org.odk.collect.android.views.MediaLayout;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
 
 public abstract class SelectWidget extends QuestionWidget {
-    private static final String SEARCH_TEXT = "search_text";
-
     protected List<SelectChoice> items;
     protected ArrayList<MediaLayout> playList;
     protected LinearLayout answerLayout;
-    protected EditText searchStr;
     private int playcounter = 0;
 
     public SelectWidget(Context context, FormEntryPrompt prompt) {
@@ -114,14 +101,6 @@ public abstract class SelectWidget extends QuestionWidget {
         super.playAllPromptText();
     }
 
-    @Override
-    protected void saveState() {
-        super.saveState();
-        if (searchStr != null) {
-            getState().putString(SEARCH_TEXT + getFormEntryPrompt().getIndex(), searchStr.getText().toString());
-        }
-    }
-
     private void playNextSelectItem() {
         if (isShown()) {
             // if there's more, set up to play the next item
@@ -176,87 +155,5 @@ public abstract class SelectWidget extends QuestionWidget {
         }
 
         return mediaLayout;
-    }
-
-    protected void doSearch(String searchStr) {
-        // First check if there is nothing on search
-        if (searchStr == null || searchStr.trim().length() == 0) {
-            createFilteredOptions(items, null);
-        } else { // Create a List with items that are relevant to the search text
-            List<SelectChoice> searchedItems = new ArrayList<>();
-            List<Integer> tagList = new ArrayList<>();
-            searchStr = searchStr.toLowerCase(Locale.US);
-            for (int i = 0; i < items.size(); i++) {
-                String choiceText = getFormEntryPrompt().getSelectChoiceText(items.get(i)).toLowerCase(Locale.US);
-                if (choiceText.contains(searchStr)) {
-                    searchedItems.add(items.get(i));
-                    tagList.add(i);
-                }
-            }
-            createFilteredOptions(searchedItems, tagList);
-        }
-    }
-
-    private void setupChangeListener() {
-        searchStr.addTextChangedListener(new TextWatcher() {
-            private String oldText = "";
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                if (!s.toString().equals(oldText)) {
-                    doSearch(s.toString());
-                }
-            }
-
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-                oldText = s.toString();
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-            }
-        });
-    }
-
-    protected void setUpSearchBox() {
-        searchStr = new EditText(getContext());
-        searchStr.setId(ViewIds.generateViewId());
-        searchStr.setTextSize(TypedValue.COMPLEX_UNIT_DIP, getAnswerFontSize());
-
-        TableLayout.LayoutParams params = new TableLayout.LayoutParams();
-        params.setMargins(7, 5, 7, 5);
-        searchStr.setLayoutParams(params);
-        setupChangeListener();
-        addAnswerView(searchStr);
-
-        String searchText = null;
-        if (getState() != null) {
-            searchText = getState().getString(SEARCH_TEXT + getFormEntryPrompt().getIndex());
-        }
-        if (searchText != null && !searchText.isEmpty()) {
-            searchStr.setText(searchText);
-            Selection.setSelection(searchStr.getText(), searchStr.getText().toString().length());
-        } else {
-            doSearch("");
-        }
-    }
-
-    private void createFilteredOptions(List<SelectChoice> searchedItems, List<Integer> tagList) {
-        removeView(answerLayout);
-        answerLayout.removeAllViews();
-
-        if (searchedItems != null && !searchedItems.isEmpty()) {
-            addButtonsToLayout(tagList);
-        }
-
-        RelativeLayout.LayoutParams params = new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT);
-        params.addRule(RelativeLayout.ALIGN_PARENT_LEFT, RelativeLayout.TRUE);
-        params.addRule(RelativeLayout.BELOW, searchStr.getId());
-        params.setMargins(10, 0, 10, 0);
-        addView(answerLayout, params);
-    }
-
-    protected void addButtonsToLayout(List<Integer> tagList) {
     }
 }
